Catch problematic warning during pnbd optimization, retry
While investigating convergence issues for the lifetimes pnbd, I found that when the following warning hits:

RuntimeWarning: invalid value encountered in subtract
  numpy.max(numpy.abs(fsim[0] - fsim[1:])) <= fatol):

it implies the likelihood is infinite (or in some cases nan) which causes the optimization to get stuck at the current parameter values.

This update catches the numpy warning as an actual RuntimeWarning, finds the individual entries that caused the likelihood to be infinite, and retries optimization with those entries removed.